% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{add_marginals}
\alias{add_marginals}
\title{Add marginals
Add requested marginals to the table}
\usage{
add_marginals(tab, tabn, row.margin.format, col.margin.format, margins)
}
\description{
Add marginals
Add requested marginals to the table
}
