% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{addhealth}
\alias{addhealth}
\title{Add Healths.}
\format{An object of class \code{data.frame} with 4016 rows and 16 columns.}
\usage{
addhealth
}
\description{
Extract from public use add health file
}
\keyword{datasets}

