% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/createClasses.R
\name{assignCommasClass}
\alias{assignCommasClass}
\title{Assign Commas Classes Function}
\usage{
assignCommasClass(variable)
}
\description{
This function creates and assigns a custom class to handle numbers with commas.
}
\examples{
assignCommasClass_function()
}
\keyword{class}
\keyword{colClasses}

