% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_educ_poverty_data.R
\name{create_educ_poverty_data}
\alias{create_educ_poverty_data}
\title{Extract and merge data}
\usage{
create_educ_poverty_data()
}
\description{
This function creates a county level education and poverty data set.
}
\keyword{data}
\keyword{education}
\keyword{poverty}

