% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Function for creating a crosstabs}
\usage{
crosstab(data, formula = NULL, row.vars = "", col.vars = "",
  format = "freq", useNA = "ifany", title = "",
  row.margin.format = "none", col.margin.format = "freq", dnames = "",
  pretty.print = FALSE)
}
\arguments{
\item{data}{Data frame to be analyzed}

\item{formula}{A formula for the table with a single row column on the left and column variables in the right}

\item{row.vars}{A single or vector of row variables, quoted. These are appended to any row variable from formula.}

\item{col.vars}{A single or vector of column variables, quoted. These are appended to any column variables from formula.}

\item{format}{"freq", "col_percent", "row_percent", "total_percent"}

\item{title}{string Title to be used for printing}

\item{row.margin.format}{"percent" to show row percents, "freq" to show row frequency, "none" for no margin}

\item{col.margin.format}{"percent" to show percents, "freq" to show column frequencies, "none" for no margin}

\item{dnames}{vector Dimension names to replace the variable names.}

\item{useNa}{"ifany", "no", "always"}

\item{pretty.pring}{logical TRUE means to add linebreaks for printing with Pander or another package.}
}
\description{
This function creates crosstabs in a flexible manner.
}
\keyword{crosstab}
