% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education.counties}
\alias{education.counties}
\title{Education counties.}
\format{An object of class \code{data.frame} with 3230 rows and 48 columns.}
\usage{
education.counties
}
\description{
County level education data
}
\keyword{datasets}

