% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education.states}
\alias{education.states}
\title{Education states.}
\format{An object of class \code{data.frame} with 52 rows and 48 columns.}
\usage{
education.states
}
\description{
States level education statistics
}
\keyword{datasets}

