% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequency}
\alias{frequency}
\title{Frequency}
\usage{
frequency(Values, freq = TRUE, percent = TRUE, cumulative.percent = FALSE,
  cumulative.freq = FALSE, useNA = "no", title = "")
}
\arguments{
\item{Values}{The variable data for which the table will be created}

\item{freq}{logical indicating if the freqencies should be displayed}

\item{percent}{logical indicating if the percents should be displayed}

\item{cumulative.percent}{logical indicating if the cumulative percents should be displayed}

\item{cumulative.freq}{logical indicating if the cumulative frequencies should be displayed}

\item{useNA}{Handling of NAs, choices are "no", "ifany", "always". Defaults to no.}

\item{title}{Optional title, enclosed in quotes. Defaults to empty.}
}
\description{
This function creates a single variable frequency table
}
\keyword{count}
\keyword{data}
\keyword{frequency}
\keyword{percent}
