% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_label.R
\name{get_col_label}
\alias{get_col_label}
\title{get_col_label}
\usage{
get_col_label(colinfo, var)
}
\description{
This function extracts a label from the colinfo matrix #' @param df  Dataframe containing columns with the col.label attribute
}
\keyword{data}
\keyword{labels}
