% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_info.R
\name{get_column_info}
\alias{get_column_info}
\title{get_col_info}
\usage{
get_column_info(df)
}
\arguments{
\item{df}{Dataframe containing columns with the col.label attribute}
}
\description{
This function creates a 2xc matrix of variable names and labels
}
\keyword{data}
\keyword{labels}
