% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggvis_percent.R
\name{ggvis_percent}
\alias{ggvis_percent}
\title{Layer Percent}
\usage{
ggvis_percent(x, x_var)
}
\description{
This function uses ggvis to create a histogram with percents
}
\examples{
ggvis_percent_function()
}
\keyword{ggvis}
\keyword{histogram}
\keyword{percent}

