% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lehman_data.R
\name{load_lehman_data}
\alias{load_lehman_data}
\title{Function for reading data from AWS S3 storage}
\usage{
load_lehman_data(df)
}
\description{
This function reads an R data set from storage. It requires that you have keys and the name of
 the bucket set as envrionmental variables. It is a convenience function for s3load() from the aws.s3
  package. AWS_DEFAULT_BUCKET is the name of the global variable holding the bucketname.
}
\keyword{aws}
\keyword{data}
\keyword{request}
