% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePercentTable.R
\name{makePercentTable}
\alias{makePercentTable}
\title{Function for creating a crosstabs}
\usage{
makePercentTable(tbl, type, row.vars, col.vars, percentages)
}
\description{
Internal function that creates percentage or proportion tables crosstabs in a flexible manner.
}
\keyword{crosstab,}
\keyword{pecents}
