% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltHMDdata.R
\name{meltHMDdata}
\alias{meltHMDdata}
\title{Function preprocessing mortality.org for making population pyramids}
\usage{
meltHMDdata(df, year)
}
\description{
This function calculates the mode for a dataframe, supporting multiple data types and multiple modes.
}
\keyword{melt}
\keyword{mortality}
\keyword{preprocess}

