% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MODE.R
\name{MODE}
\alias{MODE}
\title{Function for calculating the mode}
\usage{
MODE(dataframe)
}
\description{
This function calculates the mode for a dataframe, supporting multiple data types and multiple modes.
}
\examples{
MODE(InsectSprays)
MODE(InsectSprays$count)
based on http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode
}
\keyword{class}
\keyword{mode}

