% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_swirl_completed.R
\name{my_swirl_completed}
\alias{my_swirl_completed}
\title{Function for showing progress in swirl}
\usage{
my_swirl_completed()
}
\description{
Function for showing progress in swirl
}
\keyword{swirl}

