% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{persistentpoverty.county}
\alias{persistentpoverty.county}
\title{Persistent poverty county data.}
\format{An object of class \code{data.frame} with 3143 rows and 6 columns.}
\usage{
persistentpoverty.county
}
\description{
Count level persistent poverty and urbanization data
}
\keyword{datasets}

