% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationPyramid.R
\name{populationPyramid}
\alias{populationPyramid}
\title{Function for creating a population pyramid}
\usage{
populationPyramid(df, year, entityname)
}
\description{
This function uses ggplot to make a population pyramid from mortality.org data.
}
\keyword{graph}
\keyword{population}
\keyword{pyramid}

