% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poverty.counties}
\alias{poverty.counties}
\title{Poverty counties.}
\format{An object of class \code{data.frame} with 3143 rows and 34 columns.}
\usage{
poverty.counties
}
\description{
County level income and poverty statistics
}
\keyword{datasets}

