% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poverty.states}
\alias{poverty.states}
\title{Poverty states.}
\format{A data frame with five variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year).}
\usage{
poverty.states
}
\description{
States level income and poverty statistics
}
\keyword{datasets}

