% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{preprocess_multidimensional_tables}
\alias{preprocess_multidimensional_tables}
\title{Preprocess multidimensional tables
This converts multidimensional tables to two dimensions}
\usage{
preprocess_multidimensional_tables(tab, row.vars, col.vars, pretty.print, nvars)
}
\description{
Preprocess multidimensional tables
This converts multidimensional tables to two dimensions
}
