% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_freq.R
\name{pretty_freq}
\alias{pretty_freq}
\title{Print a frequency nicely in PDF or HTML}
\usage{
pretty_freq(freqtable)
}
\arguments{
\item{freqtable}{frequency table object (as produced by lehmansociology::frequency)}
}
\description{
Print a frequency nicely in PDF or HTML
}
\keyword{frequency,}
\keyword{print}
