% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_tab.R
\name{pretty_tab}
\alias{pretty_tab}
\title{Print a crosstab nicely in PDF or HTML}
\usage{
pretty_tab(ctab)
}
\arguments{
\item{ctab}{crossstab object as produced by lehmansociology::crosstab}
}
\description{
Print a crosstab nicely in PDF or HTML
}
\keyword{crosstab,}
\keyword{print}
