% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.frequencytable.R
\name{print.frequencytable}
\alias{print.frequencytable}
\title{Print Frequencytable results from frequency()}
\usage{
\method{print}{frequencytable}(x, ..., digits = NULL, quote = FALSE,
  right = FALSE, row.names = FALSE)
}
\arguments{
\item{x}{The object to be printed.}

\item{quote}{indicates if the string values should be returned enclosed in quotation marks.}

\item{right}{Is a logical indicating if the values should be right aligned.}

\item{row.names}{indicates if the row.names or numbers should be printed.}
}
\description{
This function controls printing for class frequencytable. It is the
same as print.data.frame except that row.names defaults to FALSE and
right defaults to FALSE. It also adds a title if this
was created when the frequencytable was created.
}
\keyword{count}
\keyword{data}
\keyword{frequency}
\keyword{percent}
