% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readmydata.R
\name{readMyData}
\alias{readMyData}
\title{Read My Data Function}
\usage{
readMyData(datasetname)
}
\description{
This function reads a stored rds dataset, harcoded to the data folder
}
\examples{
readMyDatas_function()
}
\keyword{data}
\keyword{read}

