% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacecommas.R
\name{replaceCommas}
\alias{replaceCommas}
\title{Replace Commas Function}
\usage{
replaceCommas(x)
}
\description{
This function converts a character representation of a number that contains a comma separator with a numeric value.
}
\keyword{data}
\keyword{read}
