% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp.R
\name{sdp}
\alias{sdp}
\title{The "population" standard deviation}
\usage{
sdp(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector for which a variance can be calculated}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Calculates the standarddeviation with n as the denominator.
}
\keyword{deviation,}
\keyword{population}
\keyword{standard}
