% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.crosstab.R
\name{summary.crosstab}
\alias{summary.crosstab}
\title{Summary for crosstab results from crosstab()}
\usage{
\method{summary}{crosstab}(x)
}
\arguments{
\item{x}{The crosstab object to be summarized.}
}
\description{
This function uses the summary.table function to get a summary
of a crosstab object
}
\keyword{chisquare}
\keyword{count}
\keyword{data}
\keyword{frequency}
\keyword{percent}
