% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variation.R
\name{variation}
\alias{variation}
\title{Variation Function}
\usage{
variation(x)
}
\description{
This function calculates a number of measures of variation for numeric and integer variables and returns them as a table.
Note that this uses the unbiased estimates of variance and standard deviation.
}
\examples{
variation_function(iris$Sepal.Petal)

}
\keyword{variation}

