% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varp.R
\name{varp}
\alias{varp}
\title{The "population" variance}
\usage{
varp(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{A vector for which a variance can be calculated}

\item{y}{NULL (default) or a vector, matrix or data frame
with compatible dimensions to x. The default is equivalent to y = x (but more efficient).}

\item{na.rm}{logical. Should missing values be removed?}

\item{use}{an optional character string giving a method for computing covariances in the presence of
missing values. This must be (an abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs".}
}
\description{
Calculates the variance with n as the denominator.
}
\keyword{population}
\keyword{variance,}
