% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODE.R
\name{MODE}
\alias{MODE}
\title{Function for calculating the mode}
\usage{
MODE(dataframe, Freq)
}
\description{
This function calculates the mode for a dataframe, supporting multiple data types and multiple modes.
}
\examples{
 MODE(mtcars)
 MODE(mtcars$hp)
}
\keyword{class}
\keyword{mode}

