% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readmydata.R
\name{readMyData}
\alias{readMyData}
\title{Layer Percent}
\usage{
readMyData(datasetname)
}
\description{
This function uses ggvis to create a histogram with percents
}
\examples{
layer_percent_function()
}
\keyword{ggvis}
\keyword{histogram}
\keyword{percent}

