% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variation.R
\name{variation}
\alias{variation}
\title{Variation Function}
\usage{
variation(x)
}
\description{
This function calculates a number of measures of variation for numeric and integer variables and returns them as a table.
}
\examples{
variation_function(iris$Sepal.Petal)
}
\keyword{variation}

